Less global statics in expressions
