+++ src/core/expression/qgsexpression.cpp
@@ -24,89 +24,11 @@
 #include "qgsgeometry.h"
 #include "qgsproject.h"
 #include "qgsexpressioncontextutils.h"
+#include "qgsexpression_p.h"
 
 // from parser
 extern QgsExpressionNode *parseExpression( const QString &str, QString &parserErrorMsg, QList<QgsExpression::ParserError> &parserErrors );
 
-
-struct HelpArg
-{
-  HelpArg( const QString &arg, const QString &desc, bool descOnly = false, bool syntaxOnly = false,
-           bool optional = false, const QString &defaultVal = QString() )
-    : mArg( arg )
-    , mDescription( desc )
-    , mDescOnly( descOnly )
-    , mSyntaxOnly( syntaxOnly )
-    , mOptional( optional )
-    , mDefaultVal( defaultVal )
-  {}
-
-  QString mArg;
-  QString mDescription;
-  bool mDescOnly;
-  bool mSyntaxOnly;
-  bool mOptional;
-  QString mDefaultVal;
-};
-
-struct HelpExample
-{
-  HelpExample( const QString &expression, const QString &returns, const QString &note = QString() )
-    : mExpression( expression )
-    , mReturns( returns )
-    , mNote( note )
-  {}
-
-  QString mExpression;
-  QString mReturns;
-  QString mNote;
-};
-
-
-struct HelpVariant
-{
-  HelpVariant( const QString &name, const QString &description,
-               const QList<HelpArg> &arguments = QList<HelpArg>(),
-               bool variableLenArguments = false,
-               const QList<HelpExample> &examples = QList<HelpExample>(),
-               const QString &notes = QString() )
-    : mName( name )
-    , mDescription( description )
-    , mArguments( arguments )
-    , mVariableLenArguments( variableLenArguments )
-    , mExamples( examples )
-    , mNotes( notes )
-  {}
-
-  QString mName;
-  QString mDescription;
-  QList<HelpArg> mArguments;
-  bool mVariableLenArguments;
-  QList<HelpExample> mExamples;
-  QString mNotes;
-};
-
-
-struct Help
-{
-  //! Constructor for expression help
-  Help() = default;
-
-  Help( const QString &name, const QString &type, const QString &description, const QList<HelpVariant> &variants )
-    : mName( name )
-    , mType( type )
-    , mDescription( description )
-    , mVariants( variants )
-  {}
-
-  QString mName;
-  QString mType;
-  QString mDescription;
-  QList<HelpVariant> mVariants;
-};
-
-typedef QHash<QString, Help> HelpTextHash;
-Q_GLOBAL_STATIC( HelpTextHash, sFunctionHelpTexts )
 Q_GLOBAL_STATIC( QgsStringMap, sVariableHelpTexts )
 Q_GLOBAL_STATIC( QgsStringMap, sGroups )
